Got the player moving correctly, maze2
I got my player to move on it’s own without being glitchy.  I also
Built in a counter to separate how fast the dot movies and when you con
press the buttons.  I also started my second maze and made it so if the
dot touches the wall it returns back to the start.